--- a/be/src/olap/delta_writer.cpp
+++ b/be/src/olap/delta_writer.cpp
@@ -267,6 +267,9 @@ Status DeltaWriter::write(const vectorized::Block* block, const std::vector<int>
 }
 
 Status DeltaWriter::_flush_memtable_async() {
+    if (_mem_table->empty()) {
+        return Status::OK();
+    }
     _mem_table->assign_segment_id();
     return _flush_token->submit(std::move(_mem_table));
 }