--- a/be/src/olap/memtable.h
+++ b/be/src/olap/memtable.h
@@ -146,7 +146,7 @@ public:
         _delta_writer_callback = callback;
     }
 
-    void assign_seq_id();
+    void assign_segment_id();
 
 private:
     Status _do_flush();
@@ -228,7 +228,7 @@ private:
     void _put_into_output(vectorized::Block& in_block);
     bool _is_first_insertion;
     std::function<void(MemTableStat&)> _delta_writer_callback;
-    std::optional<int32_t> _seq_id = std::nullopt;
+    std::optional<int32_t> _segment_id = std::nullopt;
 
     void _init_agg_functions(const vectorized::Block* block);
     std::vector<vectorized::AggregateFunctionPtr> _agg_functions;