… be/src/olap/rowset/vertical_beta_rowset_writer.cpp